+++ test/verify/yosupo-k-shortest-walk.test.cpp
@@ -2,12 +2,12 @@
 
 #include "../../template/template.cpp"
 
-#include "../../graph/graph-template.cpp"
-#include "../../graph/shortest-path/dijkstra.cpp"
+#include "../../graph/graph-template.hpp"
+#include "../../graph/shortest-path/dijkstra.hpp"
 
 #include "../../structure/heap/leftist-heap.cpp"
 #include "../../structure/heap/persistent-leftist-heap.cpp"
-#include "../../graph/shortest-path/k-shortest-walk.cpp"
+#include "../../graph/shortest-path/k-shortest-walk.hpp"
 
 int main() {
   int N, M, S, T, K;
@@ -16,7 +16,7 @@ int main() {
   g.read(M, 0, true, true);
   auto ret = k_shortest_walk(g, S, T, K);
   for(int i = 0; i < K; i++) {
-    if(i >= ret.size()) cout << -1 << "\n";
+    if(i >= (int)ret.size()) cout << -1 << "\n";
     else cout << ret[i] << "\n";
   }
 }
